Pipeline overhaul for multilang support
TSC in plaintext are now compiled as part of the build process
tscomp and tscomp-ja have been combined to one, and supports 7 languages
The 5 new languages require Sega SSF support
Everything required to play in EN+JA are still <4MB
Char tiles used by new languages shoved into the window plane
I did the Tanglewood blue spheres thing

+++ inc/string.h
@@ -17,6 +17,6 @@ typedef __gnuc_va_list va_list;
                        - ((sizeof (TYPE) < __va_rounded_size (char)     \
                            ? sizeof (TYPE) : __va_rounded_size (TYPE))))))
 
-uint32_t strlen(const char *str);
+uint16_t strlen(const char *str);
 uint16_t strnlen(const char *str, uint16_t maxlen);
 uint16_t sprintf(char *buffer,const char *fmt, ...) __attribute__ ((format (printf, 2, 3)));
